+++ firebase-common/src/test/java/com/google/firebase/DataCollectionPreNDefaultEnabledTest.java
@@ -27,7 +27,7 @@ import org.junit.runner.RunWith;
 import org.robolectric.annotation.Config;
 
 @RunWith(AndroidJUnit4.class)
-@Config(sdk = 19)
+@Config(sdk = 21)
 public class DataCollectionPreNDefaultEnabledTest {
 
   @Test