+++ firebase-inappmessaging/src/test/java/com/google/firebase/inappmessaging/internal/AnalyticsEventsManagerTest.java
@@ -16,8 +16,8 @@ package com.google.firebase.inappmessaging.internal;
 
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.firebase.inappmessaging.CommonTypesProto.Trigger.ON_FOREGROUND;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.eq;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
 
 import android.util.Log;